requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
db-dtypes>=1.1.1
pandas>=2.0.0
matplotlib>=3.7.0
pyarrow>=12.0.0
